Change way class methods are called
Originally, there was one main FlightPlanDB class, and the other classes
were called as attributes of it. Every time the attribute was called, it
initialised a class which inherited the key and base_url from the main
class, so that methods of that class could do authenticated requests.

In this commit, that changes. Each subclass now contains its own
methods. It inherits the url from the main class, and the key is passed
in for every method. This is slightly more annoying for the user (they
can no longer pass in the key at a single class instance, but must do so
for every library call) but it allows vastly simpler module architecture,
which is necessary to improve test coverage.

Unfortunately, all tests now fail. This will be addressed in future
commits.